resources and controllers

+++ database/migrations/2023_10_11_103101_create_order_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('book_id');
             $table->integer('quantity');
-            $table->integer('price_total');
+            $table->float('price_total');
             $table->timestamps();
         });
     }
